Move profile picture update action into profile edit page
