+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/ProfileFragment.kt
@@ -1,33 +1,23 @@
 package com.bounswe2020group3.paperlayer.profile
 
-import android.app.Activity.RESULT_OK
+
 import android.content.Context
-import android.content.Intent
-import android.os.Build
 import android.os.Bundle
-import android.provider.MediaStore
-import android.provider.OpenableColumns
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.annotation.RequiresApi
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
 import com.bounswe2020group3.paperlayer.MainActivity
 import com.bounswe2020group3.paperlayer.R
-import com.bounswe2020group3.paperlayer.data.user.Profile
 import com.bounswe2020group3.paperlayer.data.user.User
 import com.bounswe2020group3.paperlayer.profile.follow.FollowType
 import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.fragment_profile.*
 import kotlinx.android.synthetic.main.fragment_profile.imageViewProfileAvatar
 import kotlinx.android.synthetic.main.fragment_profile.layoutProfileDetail
-import kotlinx.android.synthetic.main.fragment_profile_edit.*
-import okhttp3.MediaType
-import java.io.File
-import java.io.FileOutputStream
 import javax.inject.Inject
 
 
@@ -35,10 +25,6 @@ class ProfileFragment : Fragment(), ProfileContract.View {
 
     @Inject
     lateinit var presenter: ProfileContract.Presenter
-
-    private var selectPhoto = 1
-
-    private var profile: Profile? = null
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
@@ -76,45 +62,6 @@ class ProfileFragment : Fragment(), ProfileContract.View {
         layoutFollowRequests.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigateToFollowListFromProfile, followRequestBundle)
         }
-
-        // Update profile picture
-        imageViewProfileAvatar.setOnClickListener {
-            if (this.profile != null) {
-                val intent: Intent = Intent(Intent.ACTION_PICK)
-                intent.type = "image/*"
-                startActivityForResult(intent, selectPhoto)
-            }
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (profile != null && requestCode == selectPhoto && resultCode == RESULT_OK && data != null && data.data != null) {
-            try {
-                val activity = requireActivity()
-                val uri = data.data!!
-                val contentResolver = activity.contentResolver
-                val imageType = contentResolver.getType(uri)!!
-                val mediaType = MediaType.parse(imageType)!!
-
-                // Create a temporary path name to the file
-                val inputStream = contentResolver.openInputStream(uri)!!
-                var pathName = ""
-                val cursor = contentResolver.query(uri, null, null, null)
-                cursor?.use {
-                    it.moveToFirst()
-                    pathName = cursor.getString(it.getColumnIndex(OpenableColumns.DISPLAY_NAME))
-                }
-                val file = File(activity.cacheDir, pathName)
-                val outputStream = FileOutputStream(file)
-                inputStream.copyTo(outputStream)
-
-                presenter.updateProfilePicture(profile!!.id, file, mediaType)
-            } catch (e: Exception) {
-                e.printStackTrace()
-            }
-        }
     }
 
     override fun onResume() {
@@ -130,7 +77,6 @@ class ProfileFragment : Fragment(), ProfileContract.View {
     override fun updateProfileUI(user: User) {
         try {
             val profile = user.profile.first()
-            this.profile = profile
             val fullName = "${profile.name} ${profile.lastName}"
 
             textViewEmail.text = user.email
