+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/edit/ProfileEditContract.kt
@@ -3,6 +3,8 @@ package com.bounswe2020group3.paperlayer.profile.edit
 import com.bounswe2020group3.paperlayer.mvp.Mvp
 import com.bounswe2020group3.paperlayer.data.user.Profile
 import com.bounswe2020group3.paperlayer.data.user.User
+import okhttp3.MediaType
+import java.io.File
 
 interface ProfileEditContract {
     interface Presenter : Mvp.Presenter<View> {
@@ -10,6 +12,7 @@ interface ProfileEditContract {
         fun loadUser()
         fun updateProfile(updatedProfile: Profile)
         fun getUserData(): User?
+        fun updateProfilePicture(profileId: Int, file: File, type: MediaType)
     }
 
     interface View : Mvp.View {
