+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/edit/ProfileEditPresenter.kt
@@ -6,6 +6,8 @@ import com.bounswe2020group3.paperlayer.profile.ProfileContract
 import com.bounswe2020group3.paperlayer.data.user.Profile
 import com.bounswe2020group3.paperlayer.data.user.User
 import io.reactivex.disposables.CompositeDisposable
+import okhttp3.MediaType
+import java.io.File
 import javax.inject.Inject
 
 class ProfileEditPresenter @Inject constructor(private var model: ProfileContract.Model) : BasePresenter<ProfileEditContract.View>(), ProfileEditContract.Presenter {
@@ -28,7 +30,6 @@ class ProfileEditPresenter @Inject constructor(private var model: ProfileContrac
     }
 
     override fun loadUser() {
-        Log.d("Dagger", "Profile Presenter: $model")
         view?.showLoading()
         try {
             val fetchSub = model.fetchAuthUser().subscribe(
@@ -64,5 +65,25 @@ class ProfileEditPresenter @Inject constructor(private var model: ProfileContrac
                 }
         )
         disposable.add(getProfileObservable)
+    }
+
+    override fun updateProfilePicture(profileId: Int, file: File, type: MediaType) {
+        view?.showLoading()
+        try {
+            val fetchSub = model.updateProfilePicture(profileId, file, type).subscribe(
+                    {
+                        view?.hideLoading()
+                        view?.showInfoToast("Profile updated successfully")
+                        loadUser()
+                    },
+                    {
+                        view?.hideLoading()
+                        view?.showErrorToast("An error occurred while updating the profile picture. Please try again.")
+                    }
+            )
+            disposable.add(fetchSub)
+        } catch (e: Exception) {
+            view?.showErrorToast("Please log in first.")
+        }
     }
 }
